+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.1-alpha.6",
-  "contentSha256_12": "4aaf6804e8ad",
+  "pkgVersion": "1.1.1-alpha.7",
+  "contentSha256_12": "da1d1468a19d",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.1-alpha.6.js"
+  "hacsFile": "nodalia-cards-1.1.1-alpha.7.js"
 };
-export const pkgVersion = "1.1.1-alpha.6";
-export const contentSha256_12 = "4aaf6804e8ad";
+export const pkgVersion = "1.1.1-alpha.7";
+export const contentSha256_12 = "da1d1468a19d";
 export const file = "nodalia-cards.bundle.js";